Clean up Grub configuration and add a "safe graphics" setting
It seems silly that the autoinstall command line settings end up in the
Grub configuration of the installed system, and that will probably be
fixed some day on the Ubuntu side. Hopefully this is robust enough not
to break then.

The system I'm testing with has display troubles when running without
'nomodeset', so add that at least for now.