workflow: build deb packages and an apt repository
This automates building *.deb packages from the ytl-linux-customize/
subdirectory and making an apt repository of them and any others included
in the debs/ subdirectory.

The apt repository is made available at

  deb https://linux.abitti.fi/deb ytl-linux main

and the signing key is currently at

  https://linux.abitti.fi/apt-signing-key.pub

TODO: the S3 objects are cached for 24 hours by CloudFront, cache
invalidation is badly needed for the Packages and Release files etc.

TODO: make the installation image trust this repository and install
the ytl-linux-customize package.

TODO: possibly reorganize the file hierarchy a bit so that
the apt repository stuff and the iso image stuff are better
separated.

TODO: document once this stabilizes.