Make TSO off operation also to devices starting with "eth" prefix
According to the empirical study the Ubuntu appears to name ethernet devices
with "the new" naming scheme starting with "eno" instead of legacy "eth"
prefix.

As we're unsure whether this is the case, it appears to be a good idea to make
the TSO off setting to devices starting with eth, too.