Fix installation on Linux with Firefox snap
The firefox.desktop description file exists only on Debian Linuxes where
Firefox was installed from .deb file. By default the Ubuntu uses Firefox
from snap. In this case the .desktop file is located in the snap
filesystem and for some unknown reason xdg-settings finds it with
peculiar name "firefox_firefox.desktop". This was reverse engineered by
setting the default browser with GUI tool and reading the value with
xdg-settings get default-web-browser.

In su command the -l switch creates the session with target users login
shell and in this case makes sure it is the target user's
~/.config/mimeapps.list which is edited.

The WSL command was left as-is as it appears to work.