Add boot option for networks with missing DHCP response
For some unknown reason the Ubuntu autoinstall script fails to communicate
network settings with DHCP server. As a cure issue #29 suggests
using kernel `ip=dhcp` setting for this.

Since the current method works in most environments this was not added
to the default boot option. Instead, an alternative boot option was added
with the setting.

Since we don't have an environment where the problem occurs, the effect
of the setting is unknown. Feedback is welcomed.