Do not remove CUPS client & common
Cinnamon applet "printers@cinnamon.org" requires lpstat (and other lp-tools)
to work. If these are missing, a system error is raised asking user to
send error report to Canonical.

These tools are provided by "cups-client" which requires "cups-common".
They are not providing CUPS server which is still removed if present.