Init school user in user-data and add to Docker group, remove meb user
The `meb` user has not been used in a long time and is likely not relevant anymore, so it should be fine to remove.

Also temporarily install ytl-linux-digabi2 and ytl-linux-digabi2-examnet packages in autoinstall directly, as this is the easiest way to test. These will move into being dependents of ytl-linux-customize-24 after testing.